+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
 
-<script src="js/main.js"></script>
+<script type= "module" src="js/main.js"></script>
 </body>
 </html>
